Feature buyer s1 merge to main

+++ src/main/java/com/spare4fun/core/controller/OfferController.java
@@ -1,26 +1,18 @@
 package com.spare4fun.core.controller;
 
 import com.spare4fun.core.dto.OfferDto;
-import com.spare4fun.core.dto.UserDto;
-import com.spare4fun.core.entity.Item;
 import com.spare4fun.core.entity.Offer;
-import com.spare4fun.core.entity.User;
+import com.spare4fun.core.exception.InvalidUserException;
 import com.spare4fun.core.service.ItemService;
 import com.spare4fun.core.service.OfferService;
-import com.spare4fun.core.service.UserAuthService;
 import com.spare4fun.core.service.UserService;
-import org.modelmapper.TypeMap;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/offer")
@@ -69,16 +61,24 @@ public class OfferController {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         String username = authentication.getName();
 
-        List<Offer> offers = offerService.getAllOffersBuyer(username);
         List<OfferDto> offerDtos = new ArrayList<>();
+        List<Offer> offers = offerService.getAllOffersBuyer(username);
 
         for (Offer offer : offers) {
             offerDtos.add(
                     OfferDto
                             .builder()
+                            .offerId(offer.getId())
+                            .price(offer.getPrice())
+                            .quantity(offer.getQuantity())
+                            .message(offer.getMessage())
+                            .itemId(offer.getItem().getId())
+                            .sellerId(offer.getSeller().getId())
+                            .buyerId(offer.getBuyer().getId())
                             .build()
             );
         }
+
         return offerDtos;
     }
 
@@ -87,16 +87,41 @@ public class OfferController {
     public OfferDto getOfferById(@PathVariable(value = "offerId") int offerId) {
 
         Offer offer = offerService.getOfferById(offerId);
-        OfferDto offerDto = OfferDto
-                                    .builder()
-//                                    .buyerId(offer.getBuyer().getId())
-                                    .build();
-        return offerDto;
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+        int currentUserId = userService.loadUserByUsername(username).get().getId();
+
+        if (currentUserId != offer.getSeller().getId() && currentUserId != offer.getBuyer().getId()) {
+            throw new InvalidUserException("You don't have the authorization to get the offer");
+        }
+        return OfferDto
+                .builder()
+                .offerId(offer.getId())
+                .price(offer.getPrice())
+                .quantity(offer.getQuantity())
+                .message(offer.getMessage())
+                .itemId(offer.getItem().getId())
+                .sellerName(offer.getSeller().getUsername())
+                .sellerId(offer.getSeller().getId())
+                .buyerName(offer.getBuyer().getUsername())
+                .buyerId(offer.getBuyer().getId())
+                .build();
     }
 
     @PostMapping("/deleteOffer/{offerId}")
-    public ResponseEntity<String> deleteOfferById(@PathVariable(value = "offerId") int offerId){
+    public void deleteOfferById(@PathVariable(value = "offerId") int offerId){
+
+        Offer offer = offerService.getOfferById(offerId);
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+        int currentUserId = userService.loadUserByUsername(username).get().getId();
+
+        //only buyer can delete the offer
+        if (currentUserId != offer.getBuyer().getId()) {
+            throw new InvalidUserException("You don't have the authorization to delete the offer");
+        }
         offerService.deleteOfferById(offerId);
-        return new ResponseEntity<String>("The offer has been successfully deleted!", HttpStatus.OK);
     }
 }
